+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 - Adds support for redeeming a code via a universal link.
 - Adds `code` and `type` to redemption events to improve debugging.
 
+### Fixes
+
+- Ensures that the UUID from the automatically assigned Superwall alias is passed as the `appAccountToken` for purchases when `identify(userId:)` has not been called.
+
 ## 4.3.5
 
 ### Fixes